Changed the project output directories.
In the three Delphi projects `CevoWin32`, `Integrated` and `cevo`, set
the output directory to `..\tmp\` and the unit output directory to
`..\tmp\unit\` so that build outputs are relative to the project directory.
Used the parent `..\` directory as a base since Delphi 4.0 has a bug
when running from a directory like `.\tmp` or `tmp`.

Also turned off map file and remote debug symbol generation and updated
the readme file with some information on building.  Changed the run
parameter for the `cevo` project to use the `CevoWin32.exe` program.