Changed AIProject settings for build output files.
Set the project for the Delphi AI kit to output the DLL into the tmp\
directory and unit files into tmp\units\AIProject\.  (The AI kit shares
a couple files with the cevo.dll project.)

Changed the path setting in the `AIProject.ai.txt` file to match the new
output location and added the required 64x64 bitmap, based on the
StdAI.bmp file shipped with the C-evo 1.2 installer.

Added the AI kit project to the makefile build.