Add the C# sample AI project to the makefile build.
Used `MSBuild.exe` to build the `AI.csproj` project.

Also modified the build recipies for the MSBuild commands to use the
path to the project being built rather than changing to the project
directory as part of the build; the MSBuild command works fine when
issued from the project root.