Add missing Delphi `.cfg` project files.
I mistakenly excluded these previously, but they contain project
settings and appear to be generated when a project is saved.

Added the `.cfg` file for each Delphi project, added each `.cfg` file as
a dependency of the appropriate target in the `Makefile`, and removed
`*.cfg` from the git ignore list.