Merge the `AppData_Saved.wxs` fragment into `Main.wxs`.
The `light.exe` WiX linker has a very limited command line length and
currently the `Makefile` works around this by using a glob to specify
the `.wixobj` files to link.  Unfortunately, this makes the build
fragile, so I'm merging the WiX fragment files into the main product
file.

Updated the `Makefile` with the new dependencies.