Added Delphi Package project for components.
The `Project\` files `Area.pas`, `ButtonBase.pas`, `ButtonA.pas`,
`ButtonB.pas`, `ButtonC.pas`, `ButtonN.pas` and `EOTButton.pas` are all
Delphi components and need to be registered with the IDE to prevent
errors when editing forms that contain them.

Created a component package project `Project\CevoComponents.dpk` to
create a component library (`.bpl`) that can be registered with Delphi.
Added the required `.rc` file and palette bitmaps for the components.

Added steps to the `Makefile` for package compilation.