Automatically install `CevoComponents` package.
Created the command script `install_component.cmd` for adding or
deleting a Delphi `.bpl` package.

Extended the build recipe for `CevoComponents.bpl` to automatically
install it when built, and to uninstall it on `clean`.

Updated the readme file with information on the `CevoComponents.dpk`
project and expanded notes on the command prompt build and
troubleshooting steps for building from the Delphi IDE.