Duplicate `Protocol.pas` file.
Move a copy of the `Protocol.pas` file into both `Project\` and
`Delphi_AI_Kit\` directories and remove the `Protocol\` directory.

Update the Delphi projects and the makefile with the change.

The motivation of moving `Protocol.pas` is to put the Delphi projects
back into their original state where each project had a copy of the
protocol definitions; on reflection, sharing a source file between the
main project and the AI kit is not really much better than the original
situation where each has a separate copy of the protocol definitions.