Add the C# AI template files to the Makefile build.
Add an `ai_template_out` variable to the Makefile build that lists all
the files from the C# AI template, and make those files a part of the
`all` build step, adding rules to copy the files from `AI_Template\` to
`tmp\AI_Template`.

Rename the `AI development manual.html` file to
`AI-development-manual.html` to avoid problems with `make` and spaces.

This is the first step in adding the C# AI template files to the
installer.