Add release builds of C# projects to the installer.
Change the C# projects for the `CevoDotNet.exe` loader and the
`Configurator.exe` program to place the "Release" configuration output
in `tmp\release`.

Change the WiX installer to include the release versions of these two
programs.

Add targets to the makefile to build these release versions.  Add the
debug versions to the `all` target so that they still get built and
placed in `tmp\`.