Change ID for the start menu group in the installer.
Change the directory ID `ProgramMenuDir` to `Dir_C_evo_x` to follow the
convention for other directories.

Remove the `Name` attribute for `ProgramFilesFolder` and
`ProgramMenuFolder` since they're system folders.

Change the registry key for the `Dir_C_evo_x` component to
`Software\C-evo-x\Installed\Start Menu Group` so the purpose of the key
is clear.