Use `c-evo-x.org` on the Start view.
Change the "On the web: c-evo.org" button on the first tab of the Start
view to use "c-evo-x.org" and to open that link instead.

Also update the hard-coded path in the code:

    AI Template\AI development manual.html

to use underscores and dashes instead of spaces:

    AI_Template\AI-development-manual.html

to match the name changes in the code needed to work with the Makefile.

Add the `.vscode` directory to the git ignore file.